+++ cli/cmd/root.go
@@ -187,6 +187,7 @@ After that tt will be able to manage the application using 'replicaset_example' 
 		NewEnvCmd(),
 		NewDownloadCmd(),
 		NewKillCmd(),
+		NewLogCmd(),
 	)
 	if err := injectCmds(rootCmd); err != nil {
 		panic(err.Error())