+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `tt log`: a module for viewing instances logs. Supported options:
+  * `--lines` number of lines to print.
+  * `--follow` print appended data as log files grow.
+
 ## [2.3.1] - 2024-06-13
 
 ### Added